Move Pigsty from Core to Data
