+++ src/Microsoft.ML.Data/StaticPipe/PipelineColumnAnalyzer.cs
@@ -2,12 +2,12 @@
 using System.Linq;
 using System.Reflection;
 using System.Runtime.CompilerServices;
-using Microsoft.ML.Core.StrongPipe.Columns;
+using Microsoft.ML.Data.StaticPipe.Runtime;
 using Microsoft.ML.Runtime;
 using System;
 using Microsoft.ML.Runtime.Internal.Utilities;
 
-namespace Microsoft.ML.Core.StrongPipe
+namespace Microsoft.ML.Data.StaticPipe
 {
     internal static class PipelineColumnAnalyzer
     {
